requirements.txt
fastapi
uvicorn
firebase-admin
python-multipart
sqlalchemy
psycopg2-binary
requests
google-generativeai
python-dotenv